+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Release History
 
+## 1.11.0 (TBD)
+
+## New Features
+
+- Added support for `RelationalGroupedDataframe.pivot()` to access `pivot` in the following pattern `Dataframe.group_by(...).pivot(...)`.
+
 ## 1.10.0 (2023-11-03)
 
 ### New Features